livekit_voice_ai/requirements.txt
# Livekit Voice AI Agent — All Dependencies
# Install with: make install

# ─── Build Tools (MUST BE FIRST) ─────────────────────────────────
setuptools>=65.0
wheel>=0.38.0
pip>=23.0
flit_core>=3.8.0

# ─── ADK & Google Cloud (REQUIRED) ───────────────────────────
google-adk>=1.28.1
google-genai>=0.3.0
google-cloud-aiplatform>=1.42.0
pydantic>=2.0

# ─── CLI & Web UI Tools (REQUIRED) ───────────────────────────
fastapi>=0.100.0
uvicorn>=0.23.0
watchdog>=3.0.0
click>=8.1.0
pyyaml>=6.0

# ─── MCP & Authentication (REQUIRED) ───────────────────────
mcp>=0.1.0
authlib>=1.2.0

# ─── Utilities (REQUIRED) ──────────────────────────────────
python-dotenv>=1.0
requests>=2.31.0

# ─── LIVEKIT VOICE STACK (with dependency workaround) ──────────────────
# NOTE: Has opentelemetry conflict - installing separately with --no-deps
livekit-agents==1.5.1
livekit-plugins-deepgram>=0.8.0
livekit-plugins-silero>=0.8.0

# ─── Livekit Dependencies (manually listed to avoid conflict) ──────────────
numpy>=1.21.0
protobuf>=3.20.0
aiohttp>=3.8.0
websockets>=10.0
av>=8.0.0
scipy>=1.7.0
soundfile>=0.10.0

# ─── ML Model Loading ─────────────────────────────────────────────────────
# Required by FunctionGemma for device mapping
accelerate>=0.24.0

# ─── FunctionGemma for local bash dispatch (REQUIRED) ──────────────
# Used by bash_executor agent for efficient tool dispatch
transformers>=4.36
torch>=2.0

# ─── ADK from GitHub (INSTALLED SEPARATELY) ─────────────────────
# This is installed via: uv pip install git+https://github.com/google/adk-python.git --no-build-isolation
# (listed here for reference; see Makefile for actual installation)
